Revert to a6f0d27

+++ Nade/src/Components/World.h
@@ -5,7 +5,7 @@ namespace Nade{
     class World{
     public:
         int skyboxType = 0;
-        float ambient = (float)0.1;
+        float ambient = 0.1;
         void Bind(unsigned int shader){
 		    Nade::Shader::Set1f(shader, "ambient", ambient);
         }
